re-adjust various places to match the recent pushes/refactoring
